refactor: use stub on web to prevent crashes during a build
