+++ src/index.ts
@@ -5,4 +5,5 @@ export * from './animated';
 export * from './replicas';
 export * from './context';
 export * from './hooks';
+export * from './constants';
 export * from './types';
