+++ src/replicas.ts
@@ -10,7 +10,8 @@ import {
 } from 'react-native-reanimated';
 import { useReanimatedKeyboardAnimation } from './hooks';
 
-import { AndroidSoftInputModes, KeyboardController } from './native';
+import { KeyboardController } from './native';
+import { AndroidSoftInputModes } from './constants';
 
 const availableOSEventType = Platform.OS === 'ios' ? 'Will' : 'Did';
